requirements.txt
aiofiles==24.1.0
annotated-types==0.7.0
blinker==1.9.0
charset-normalizer==3.4.0
click==8.1.7
filelock==3.16.1
Flask==3.1.0
fsspec==2024.10.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
Hypercorn==0.17.3
hyperframe==6.0.1
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
numpy==2.1.3
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
pillow==11.0.0
priority==2.0.0
pydantic==2.10.2
pydantic_core==2.27.1
pyhumps==3.8.0
pypugjs==5.11.0
Quart==0.19.9
quart-schema==0.20.0
quart-uploads==0.0.4
setuptools==75.6.0
six==1.16.0
sympy==1.13.1
torch==2.5.1
torchvision==0.20.1
triton==3.1.0
types-aiofiles==24.1.0.20240626
typing_extensions==4.12.2
Werkzeug==3.1.3
wsproto==1.2.0
